requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.8.1
anyio==4.12.0
APScheduler==3.11.2
asgiref==3.11.0
backoff==2.2.1
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
dnspython==2.8.0
duckduckgo_search==8.1.1
durationpy==0.10
eventlet==0.34.2
fastapi==0.127.0
filelock==3.20.1
discord.py==2.3.2
python-telegram-bot==20.7
python-multipart==0.0.6
flatbuffers==25.12.19
fsspec==2025.12.0
google-ai-generativelanguage==0.6.6
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kubernetes==34.1.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.2
numpy==1.26.2
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.6.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
overrides==7.7.0
packaging==25.0
pandas==2.1.4
pillow==12.0.0
pluggy==1.6.0
posthog==7.4.2
primp==0.15.0
proto-plus==1.27.0
protobuf==4.25.8
pulsar-client==3.8.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==2.2.2
sentencepiece==0.2.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.5.2
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.17.3
wsproto==1.3.2
zipp==3.23.0

# New autopilot integrations
tweepy>=4.14.0
linkedin-api>=2.0.0
google-auth-oauthlib>=1.0.0

# Brain Station / Knowledge Management
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Performance Monitoring
psutil>=5.9.0

# Uvicorn Production Performance
uvloop>=0.19.0; sys_platform != 'win32'
httptools>=0.6.0

# Local-First Voice (offline TTS/STT)
faster-whisper>=1.0.0
piper-tts>=1.2.0

# Image processing for Rewind feature
Pillow>=10.0.0

